Curb fmemopen to GLIBC 2.17
For reasons which are still unclear, when we link a program in CGO, it links
against the fmemopen(3) of the building machine, not the one required in
libquark_big.a (which is older).

We build libquark_big.a in centos7, which curbs the symbol to its own version
glibc version 2.2.5, but when we link it in a Go binary, the symbol is promoted
to whatever the host glibc has, oddly enough this only happens with this
specific symbol "fmemopen@GLIBC_2.2.5".

In glibc 2.22 (not to be confused with 2.2) fmemopen(3) was re-implemented since
the original was buggy, in order to guarantee we can run on old distributions,
we have to use the old buggy symbol, but fear not because we never use it, it's
something libbpf wants to link but it's not in our path.

We use ld(1) wrap magic to guarantee linking with the old symbol.
With this, mjwolf@ confirms the tests for auditbeat pass.