Commit header files by bot action.
In order to keep the required header files up-to-date with quark, these files
will be auto-committed by the buildkite bot when they change.

This also changes the commit action to only commit when (1) running on a PR build,
(2) the quark submodule has changed.  This should help minimize the size of the repo, by only
committing the large files when needed.